src/main/java/com/example/devoxx/PassengerService.java
package com.example.devoxx;

import java.util.Collection;
import java.util.List;

import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.authority.AuthorityUtils;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;

public class PassengerService {
    final static class PassengerDetails extends Passenger implements UserDetails {

        PassengerDetails(Passenger user) {
            super(user);
        }

        @Override
        public Collection<? extends GrantedAuthority> getAuthorities() {
            List<GrantedAuthority> authorities = AuthorityUtils.createAuthorityList("PASSENGER");
            if (this.isPremium()) {
                authorities.add(new SimpleGrantedAuthority("PREMIUM"));
            }
            return authorities;
        }

        @Override
        public boolean isAccountNonExpired() {
            return true;
        }

        @Override
        public boolean isAccountNonLocked() {
            return true;
        }

        @Override
        public boolean isCredentialsNonExpired() {
            return true;
        }

        @Override
        public boolean isEnabled() {
            return true;
        }
    }
}
